checkpoint moving stuff to test_HAL_ShimSQ
